requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
firebase-admin>=6.4.0
google-cloud-storage>=2.15.0
python-multipart>=0.0.9
